+++ src/eleventyComputed/page/gitCreated.ts
@@ -1,4 +1,4 @@
-const spawn = require("cross-spawn");
+import spawn from "cross-spawn";
 import path from "path";
 
 async function gitCreatedFromGit(cwd: string, fileName: string) {