+++ src/filters/menu.tsx
@@ -2,6 +2,7 @@ import Nano, { h, Fragment } from "nano-jsx";
 import path from "path";
 import { removeExtension } from "../utils/removeExtension";
 import { normalizeUrl } from "../utils/normalizeUrl";
+import { evaluatePermalink } from "../utils/evaluatePermalink";
 
 interface TocConfig {
     from: number;
@@ -114,7 +115,7 @@ const MenuItemPage = ({
 
     const isCurrentPage = currentPageStem === pageOfMenuItem.filePathStem;
 
-    const outputPath = pageOfMenuItem.data.permalink ?? pageOfMenuItem.data.page.outputPath;
+    const outputPath = evaluatePermalink(pageOfMenuItem) ?? pageOfMenuItem.data.page.outputPath;
     const pageHref = normalizeUrl(pageOfMenuItem.data.baseUrl + outputPath);
 
     // If we are on the current page, we render the TOC