+++ tests/components/navbar.test.ts
@@ -1,8 +1,9 @@
 import { expect, test } from 'vitest'
 import { formatHTML } from "../utils/_formatHTML";
-const Eleventy = require("@11ty/eleventy");
+import Eleventy from "@11ty/eleventy";
 
 test("minimal navbar is generated if no navbar.nacara data is provided", async () => {
+    console.log(process.cwd());
     const elev = new Eleventy(
         "./fixtures/navbar-0/",
         "./fixtures/navbar-0/_site",
@@ -12,7 +13,7 @@ test("minimal navbar is generated if no navbar.nacara data is provided", async (
     );
 
     const json = await elev.toJSON();
-    const formattedResult = formatHTML(json[0].content);
+    const formattedResult = await formatHTML(json[0].content);
 
     expect(formattedResult).toMatchSnapshot();
 });
@@ -27,7 +28,7 @@ test("supports simple link elements in the start section", async () => {
     );
 
     const json = await elev.toJSON();
-    const formattedResult = formatHTML(json[0].content);
+    const formattedResult = await formatHTML(json[0].content);
 
     expect(formattedResult).toMatchSnapshot();
 });
@@ -42,7 +43,7 @@ test("pinned simple link elements are always displayed on mobile", async () => {
     );
 
     const json = await elev.toJSON();
-    const formattedResult = formatHTML(json[0].content);
+    const formattedResult = await formatHTML(json[0].content);
 
     expect(formattedResult).toMatchSnapshot();
 });
@@ -57,7 +58,7 @@ test("navbar endItems are rendered using their icons on desktop only and rendere
     );
 
     const json = await elev.toJSON();
-    const formattedResult = formatHTML(json[0].content);
+    const formattedResult = await formatHTML(json[0].content);
 
     expect(formattedResult).toMatchSnapshot();
 });