+++ tests/eleventyComputed/page/absolutePath.test.ts
@@ -1,5 +1,6 @@
 import { expect, test } from 'vitest'
-const Eleventy = require("@11ty/eleventy");
+import Eleventy from "@11ty/eleventy";
+
 import { formatHTML } from "../../utils/_formatHTML";
 import path from "path";
 
@@ -13,7 +14,7 @@ test("returns the absolutePath of the file", async () => {
     );
 
     const json = await elev.toJSON();
-    const formattedResult = formatHTML(json[0].content);
+    const formattedResult = await formatHTML(json[0].content);
 
     const expected = path.resolve(
         "./fixtures/absolute-path-0/docs/getting-started/page1.md"