+++ tests/fixtures/menu-6/_data/eleventyComputed.js
@@ -1,0 +1,5 @@
+module.exports = {
+    permalink: data => {
+        return data.page.filePathStem + ".html";
+    }
+}