+++ tests/fixtures/menu-7/docs/docs.11tydata.js
@@ -1,0 +1,5 @@
+export default {
+    permalink: data => {
+        return data.page.filePathStem + ".html";
+    }
+}